requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0
idna==3.4
iniconfig==1.1.1
libretranslatepy==2.1.1
loguru==0.6.0
lxml==4.9.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
pytesseract==0.3.10
pytest==7.2.0
requests==2.28.1
tomli==2.0.1
translate==3.6.1
urllib3==1.26.12
